feat: (#46) update를 헤더에서 전체업데이트로 변경
update에서 일반 글을 받으면 그걸 헤더로 변환해서 저장하는식으로 변환. noticeBoard에 contents필드 다시 되살림. 기존 headers들 삭제하고 업데이트하는걸로 변경
