+++ src/main/java/com/example/semiwiki_backend/domain/notice_board/dto/response/NoticeBoardDetailResponseDto.java
@@ -14,6 +14,8 @@ import java.util.List;
 public class NoticeBoardDetailResponseDto {
     private String title;
 
+    private String contents;
+
     private LocalDateTime createdAt;
 
     private LocalDateTime modficatedAt;
