+++ src/main/java/com/example/semiwiki_backend/domain/user/controller/UserController.java
@@ -30,5 +30,4 @@ public class UserController {
         }
         return ResponseEntity.ok().body(noticeBoardListResponseDtos);
     }
-
 }
